creates ThreeDOFRobot and TestPoint header files
